+++ src/components/protected/customStyles.js
@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { makeStyles } from '@material-ui/core/styles';
